more cleanup code

+++ terraform/destroy.sh
@@ -1,7 +1,19 @@
 #! /bin/bash
 
+TERRAFORM_MSSQL="../config/terraform.tfvars"
+TERRAFORM_TEMPLATE="../config/terraform-template.tfvars"
+
+if [ ! -f "${TERRAFORM_MSSQL}" ]; then
+    echo "ERROR: The file ${TERRAFORM_MSSQL} is missing."
+    exit 1
+fi
+
+if [ ! -f "${TERRAFORM_TEMPLATE}" ]; then
+    echo "ERROR: The file ${TERRAFORM_TEMPLATE} is missing."
+    exit 1
+fi
+
 terraform destroy \
 -auto-approve \
---var-file=../../config/terraform.tfvars.mssql \
---var-file=../config/terraform.tfvars \
---var-file=../config/terraform-mssql.tfvars
+--var-file="${TERRAFORM_MSSQL}" \
+--var-file="${TERRAFORM_TEMPLATE}"
